refactor(admin): split StoryManagementPanel.tsx into focused modules (Related to #1944)
Extract 576L monolith into three focused components:
- StoryListTable.tsx (175L) — filter bar + sortable table + loading/error/empty states
- StoryFormModal.tsx (207L) — create/edit form modal with Field helper
- StoryManagementPanel.tsx (281L) — slim orchestrator (state + API calls only)

All 9 characterization tests pass (GREEN → refactor → GREEN).
StoryImportPanel skipped: no bulk-upload/Drive-sync code exists in current codebase.
